move CompletableFuture conversion to org.http4s.internal
