+++ core/src/main/scala/org/http4s/internal/package.scala
@@ -1,5 +1,8 @@
 package org.http4s
 
+import java.util.concurrent.{CancellationException, CompletableFuture, CompletionException}
+import java.util.function.BiFunction
+
 import cats.effect._
 import cats.implicits._
 import scala.concurrent.ExecutionContext
@@ -116,5 +119,22 @@ package object internal {
             }(direct)
           }
       }
+    }
+
+  // Adapted from https://github.com/typelevel/cats-effect/issues/160#issue-306054982
+  private[http4s] def fromCompletableFuture[F[_], A](fcf: F[CompletableFuture[A]])(
+      implicit F: Concurrent[F]): F[A] =
+    fcf.flatMap { cf =>
+      F.cancelable(cb => {
+        cf.handle[Unit](new BiFunction[A, Throwable, Unit] {
+          override def apply(result: A, err: Throwable): Unit = err match {
+            case null => cb(Right(result))
+            case _: CancellationException => ()
+            case ex: CompletionException if ex.getCause ne null => cb(Left(ex.getCause))
+            case ex => cb(Left(ex))
+          }
+        })
+        F.delay { cf.cancel(true); () }
+      })
     }
 }
